backend/requirements.txt
# FastAPI and ASGI server
fastapi>=0.95.0,<0.96.0
uvicorn>=0.22.0,<0.23.0
pydantic>=1.10.7,<2.0.0
email-validator>=2.0.0,<3.0.0

# Database
sqlalchemy>=2.0.9,<2.1.0
psycopg2-binary>=2.9.6,<3.0.0
alembic>=1.10.3,<1.11.0

# Security
python-jose>=3.3.0,<3.4.0
passlib>=1.7.4,<1.8.0
bcrypt>=4.0.1,<4.1.0
python-multipart>=0.0.6,<0.1.0

# Environment variables
python-dotenv>=1.0.0,<1.1.0

# LLM Providers (uncomment the one you'll use)
openai>=0.27.6,<0.28.0
# google-generativeai>=0.1.0,<0.2.0  # For Gemini
# anthropic>=0.5.0,<0.6.0  # For Claude
# deepseek>=0.1.0,<0.2.0  # For DeepSeek

# Utilities
requests>=2.28.2,<2.29.0
python-dateutil>=2.8.2,<2.9.0